chore: update documentation for robustness
