+++ docs/source/changelog/index.rst
@@ -1,4 +1,5 @@
-*************
+*********
 Changelog
-*************
+*********
 
+Check the change history in the `Changelog <https://github.com/holistic-ai/holisticai/blob/main/CHANGELOG.md>`_.
