+++ docs/source/conf.py
@@ -17,8 +17,7 @@
 import sys
 import os
 import shutil
-import nbformat
-from nbconvert.preprocessors import ExecutePreprocessor
+
 sys.path.insert(0, os.path.abspath('.'))
 import utils.xai_image_plots as xai_utils
 import inspect
@@ -29,48 +28,7 @@ for name, obj in inspect.getmembers(xai_utils):
     if inspect.isfunction(obj) and name.startswith('image_'):
         obj()
 
-def run_notebook(notebook_path):
-    with open(notebook_path) as f:
-        nb = nbformat.read(f, as_version=4)
-        ep = ExecutePreprocessor(timeout=600, kernel_name='python3')
-        try:
-            ep.preprocess(nb, {'metadata': {'path': os.path.dirname(notebook_path)}})
-            with open(notebook_path, 'wt') as f:
-                nbformat.write(nb, f)
-            print(f"Executed: {notebook_path}")
-            print(f"Output: {notebook_path}")
-        except Exception as e:
-            print(f"Error executing the notebook {notebook_path}: {e}")
 
-def run_all_notebooks(folder_path):
-    for root, _, files in os.walk(folder_path):
-        for file in files:
-            if file.endswith('.ipynb'):
-                notebook_path = os.path.join(root, file)
-                run_notebook(notebook_path)
-
-sys.path.insert(0, os.path.abspath(".."))
-work_dir = '/'.join(os.getcwd().split("/")[:-2])
-src_path = os.path.join(work_dir,'src')
-sys.path.insert(0, src_path)
-
-bias_tutorial_path = os.path.join(work_dir, 'tutorials/bias')
-dataset_tutorial_path = os.path.join(work_dir, 'tutorials/datasets')
-xai_tutorial_path = os.path.join(work_dir, 'tutorials/explainability')
-
-def copy_folder(origen, destino):
-    try:
-        if not os.path.exists(destino):
-            os.makedirs(destino)
-        shutil.copytree(origen, destino, dirs_exist_ok=True)
-        #run_all_notebooks(destino)
-        print(f"Folder copied from {origen} to {destino} sucessfully.")
-    except Exception as e:
-        print(f"Error when trying to copy folder: {e}")
-
-for path in [bias_tutorial_path, dataset_tutorial_path, xai_tutorial_path]:
-    dirname = os.path.basename(path)
-    copy_folder(path, os.path.join(os.getcwd(), 'gallery', 'tutorials', dirname))
 
 # -- Project information -----------------------------------------------------
 
@@ -115,6 +73,11 @@ intersphinx_mapping = {
     "pandas": ("https://pandas.pydata.org/pandas-docs/stable/", None),
     "sklearn": ("https://scikit-learn.org/stable/", None),
     "matplotlib": ("https://matplotlib.org/stable/", None)
+}
+
+source_suffix = {
+    '.rst': 'restructuredtext',
+    '.md': 'markdown',
 }
 
 nbsphinx_allow_errors = True  # Permitir errores en los notebooks
