+++ docs/source/gallery/tutorials/explainability.rst
