+++ docs/source/gallery/tutorials/robustness.rst
@@ -1,0 +1,9 @@
+Robustness
+==========
+
+Measuring
+---------
+
+Here are some tutorials where different systems are assessed using the metrics.
+
+- `Measuring Robustness Classification <robustness/measuring_robustness/measuring_robustness_classification.ipynb>`_
