+++ docs/source/getting_started/bias/metrics/multi_classification.rst
@@ -9,7 +9,7 @@ Multi-Class Classification
 Equality of Outcome Metrics
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-1. **Multiclass Statistical Parity:**  This metric computes statistical parity for a classification task with multiple classes and a protected attribute with multiple groups. For each group compute the vector of success rates for entering each class. Compute all distances (mean absolute deviation) between such vectors. Then aggregate them using the mean, or max strategy.
+**Multiclass Statistical Parity:**  This metric computes statistical parity for a classification task with multiple classes and a protected attribute with multiple groups. For each group compute the vector of success rates for entering each class. Compute all distances (mean absolute deviation) between such vectors. Then aggregate them using the mean, or max strategy.
 
     The accepted values and bounds for this metric are the same as the 1d case. A value of 0 is desired. Values below 0.1 are considered fair.
 
@@ -27,7 +27,7 @@ If the mean strategy is selected, the metric is defined as:
 Equality of Opportunity Metrics
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-1. **Multiclass Equality of Opportunity:** This metric is a multiclass generalisation of Equality of Opportunity. For each group, compute the matrix of error rates (normalised confusion matrix). Compute all distances (mean absolute deviation) between such matrices. Then aggregate them using the mean, or max strategy.
+**Multiclass Equality of Opportunity:** This metric is a multiclass generalisation of Equality of Opportunity. For each group, compute the matrix of error rates (normalised confusion matrix). Compute all distances (mean absolute deviation) between such matrices. Then aggregate them using the mean, or max strategy.
 
     The accepted values and bounds for this metric are the same as the 1d case. A value of 0 is desired. Values below 0.1 are considered fair.
 
@@ -42,7 +42,7 @@ If the mean strategy is selected, the metric is defined as:
     EOD_{mean} = \frac{1}{n} \sum_{i=1}^{n} \left| TPR_{g} - TPR_{a} \right|
 
 
-2. **Multiclass Average Odds:** This metric is a multiclass generalisation of Average Odds. For each group, compute the matrix of error rates (normalised confusion matrix). Average these matrices over rows, and compute all pariwise distance (mean absolute deviation) between the resulting vectors. Aggregate results using either mean or max strategy.
+**Multiclass Average Odds:** This metric is a multiclass generalisation of Average Odds. For each group, compute the matrix of error rates (normalised confusion matrix). Average these matrices over rows, and compute all pariwise distance (mean absolute deviation) between the resulting vectors. Aggregate results using either mean or max strategy.
 
     The accepted values and bounds for this metric are the same as the 1d case. A value of 0 is desired. Values below 0.1
     are considered fair.
@@ -57,7 +57,7 @@ If the mean strategy is selected, the metric is defined as:
 .. math::
     AOD_{mean} = \frac{1}{n} \sum_{i=1}^{n} \left| \frac{1}{2}[(TPR_{g} - TPR_{a}) + (FPR_{g} - FPR_{a})] \right|
 
-3. **Multiclass True Positive Difference:** This metric is a multiclass generalisation of TPR Difference. For each group, compute the matrix of error rates (normalised confusion matrix). Compute all distances (mean absolute deviation) between the diagonals of such matrices. Then aggregate them using the mean, or max strategy.
+**Multiclass True Positive Difference:** This metric is a multiclass generalisation of TPR Difference. For each group, compute the matrix of error rates (normalised confusion matrix). Compute all distances (mean absolute deviation) between the diagonals of such matrices. Then aggregate them using the mean, or max strategy.
 
     The accepted values and bounds for this metric are the same as the 1d case. A value of 0 is desired. Values below 0.1 are considered fair.
